+++ src/requests/Request.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Request.cpp                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emadriga <emadriga@student.42madrid.com>   +#+  +:+       +#+        */
+/*   By: jvacaris <jvacaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/30 20:13:54 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/24 15:23:33 by emadriga         ###   ########.fr       */
+/*   Updated: 2023/05/24 16:35:13 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */